Fix rendering in Oculus Mode! The previous update, which switched the Ogre renderer from legacy GL to GL3Plus, broke most rendering due to incompatibilities in many materials.
This change reverts back to the legacy GL renderer, and relies on GL 2.0
extensions (specifically GL_ARB_framebuffer_object) to do the texture
blitting from the Ogre render target to the OVR texture instead of the GL3
functions we were able to use before. Everything looks to be working well.